+++ Sixpence.ORM.Test/BaseEntityTest.cs
@@ -28,7 +28,7 @@ namespace Sixpence.ORM.Test
             SixpenceORMBuilderExtension.UseORM(null, options =>
             {
                 options.AutoGenerate = true;
-                options.EntityClassNameCase = ClassNameCase.UnderScore;
+                options.EntityClassNameCase = NameCase.Pascal;
             });
         }
 
@@ -41,26 +41,16 @@ namespace Sixpence.ORM.Test
         [Test]
         public void Check_Resolve_Entity()
         {
-            var entity = ServiceContainer.Resolve<IEntity>(className => MatchEntity.CompareEntityName(className, "user_info"));
+            var entity = ServiceContainer.Resolve<IEntity>(className => EntityCommon.CompareEntityName(className, "user_info"));
             Assert.IsNotNull(entity);
         }
 
         [Test]
         public void Check_Resolve_EntityManagerPlugin()
         {
-            var plugin = ServiceContainer.ResolveAll<IEntityManagerPlugin>(item => MatchEntity.MatchEntityManagerPlugin(item, "user_info"));
+            var plugin = ServiceContainer.ResolveAll<IEntityManagerPlugin>(item => EntityCommon.MatchEntityManagerPlugin(item, "user_info"));
             Assert.IsNotNull(plugin);
         }
-    }
-
-    [Entity("user_info", "用户信息")]
-    public class user_info : BaseEntity
-    {
-        [PrimaryColumn]
-        public string id { get; set; }
-
-        [Column("code", "编码", DataType.Varchar, 100, false)]
-        public string code { get; set; }
     }
 
     public class UserInfoPlugin : IEntityManagerPlugin