+++ Sixpence.ORM.Test/TestRepository.cs
@@ -16,6 +16,7 @@ namespace Sixpence.ORM.Test
     public class TestRepository
     {
         private Repository<Test> testRepository;
+        private Repository<TestGuidNumber> testGuidNumerRepository;
 
         [SetUp]
         public void SetUp()
@@ -28,6 +29,7 @@ namespace Sixpence.ORM.Test
             CallContext<CurrentUserModel>.SetData(CallContextType.User, new CurrentUserModel() { Id = "1", Code = "1", Name = "test" });
             SixpenceORMBuilderExtension.UseORM(null);
             testRepository = new Repository<Test>();
+            testGuidNumerRepository = new Repository<TestGuidNumber>();
         }
 
         [Test]
@@ -45,7 +47,7 @@ namespace Sixpence.ORM.Test
         {
             var dataList = testRepository.Find();
             Assert.IsTrue(dataList != null && dataList.ToList().Count > 0);
-            var data = testRepository.FindOne(new Dictionary<string, object>() { { "id", "124" } });
+            var data = testRepository.FindOne(new { id = "124" });
             Assert.IsNotNull(data);
             data = testRepository.FindOne("124");
             Assert.IsNotNull(data);
@@ -79,6 +81,19 @@ namespace Sixpence.ORM.Test
 
             testRepository.Delete(new string[] { id });
             data = testRepository.FindOne(id);
+            Assert.IsNull(data);
+        }
+
+        [Test]
+        [Order(5)]
+        public void Check_Guid_Number_Generate()
+        {
+            var id = testGuidNumerRepository.Create(new TestGuidNumber());
+            Assert.IsTrue(!string.IsNullOrEmpty(id));
+
+            testGuidNumerRepository.Delete(id);
+
+            var data = testGuidNumerRepository.FindOne(id);
             Assert.IsNull(data);
         }
     }