+++ Sixpence.ORM/Driver/DriverType.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -7,7 +8,7 @@ namespace Sixpence.ORM.Driver
 {
     public enum DriverType
     {
-        Postgresql,
-        Mysql
+        [Description("Postgres")]
+        Postgresql
     }
 }