+++ Sixpence.ORM/Driver/IDbDriver.cs
@@ -15,6 +15,8 @@ namespace Sixpence.ORM.Driver
     [ServiceRegister]
     public interface IDbDriver
     {
+        string Provider { get; }
+
         /// <summary>
         /// 获取数据库连接对象
         /// </summary>