+++ Sixpence.ORM/Entity/Interface/IPreCreateEntities.cs
@@ -1,17 +1,16 @@
 ﻿using Sixpence.Common.IoC;
-using Sixpence.ORM.Entity;
 using Sixpence.ORM.EntityManager;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Sixpence.ORM.Extensions
+namespace Sixpence.ORM.Entity
 {
     /// <summary>
-    /// 创建所有实体后
+    /// 创建所有实体前
     /// </summary>
     [ServiceRegister]
-    public interface IPostCreateEntities
+    public interface IPreCreateEntities
     {
         void Execute(IEntityManager manager, IEnumerable<IEntity> entities);
     }