+++ Sixpence.ORM/Entity/Interface/IPreCreateEntity.cs
@@ -1,11 +1,10 @@
 ﻿using Sixpence.Common.IoC;
-using Sixpence.ORM.Entity;
 using Sixpence.ORM.EntityManager;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Sixpence.ORM.Extensions
+namespace Sixpence.ORM.Entity
 {
     /// <summary>
     /// 创建实体前