+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.1.0
+
+### Features
+
++ 添加了批量操作方法：`BulkCreate`、`BulkUpdate`、`BulkCreateOrUpdate`、`BulkDelete`
++ SQL查询参数类型从`Dictionary<string, object>`变更为`object`
++ 实体`ID`支持`Number`类型
++ 特性`EntityAttribute`的属性调整为选填
+
 ## 2.0.0
 
 ### Features